move parsed parentheses tests to ast
